+++ lib/layouts/sandboxDefault.layout.php
@@ -5,7 +5,7 @@ class sandboxDefaultLayout extends waLayout
     public function execute(): void
     {
         $this->view->assign([
-            'contact_id' => wa()->getUser()->getId(),
+            'contact_id' => sandbox()->getContactId(),
         ]);
     }
 }